Purify Functions
Purify functions and separate pure from impure functions. Test writing will initially focus on pure functions and further purifying the functions that remain impure.
